Add files via upload
I am excited, about the start of my journey with CodexCue as an intern, I am grateful for the opportunity to contribute and grow alongside a team that values creativity and collaboration. Joining CodexCue is more than just an internship; it's a chance to dive into meaningful projects, expand my skill set, and make a tangible impact. Excited to immerse myself in the dynamic culture and learn from industry leaders. 
Thank you, CodexCue, for believing in my potential and providing an environment where I can thrive to make great things happen together!